Refactor reports.
This restructures everything around reports and formatters:

* Splits up all the report classes contained in report.rb into separate files
* Moves all formatters (except the code_climate one, see below) into report/formatter
* Puts all reports that are special in the sense that they consist of more than
  just a report class into a separate folder:
  * html_report (because of the template)
  * code_climate (because the formatter is 100% code_climate specific so it should not
    end up in report/formatter and then there is also the yml configuration)
* Given a "foo" directory with the files "bar.rb" and "baz.rb" there will always be a
  "foo.rb" on the same level as the directory which will require "bar" and "baz"
