+++ lib/reek.rb
@@ -4,4 +4,4 @@
 #
 require_relative 'reek/version'
 require_relative 'reek/examiner'
-require_relative 'reek/report/report'
+require_relative 'reek/report'
