+++ lib/reek/report/code_climate.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+require_relative 'code_climate/code_climate_formatter'
+require_relative 'code_climate/code_climate_report'
