+++ lib/reek/report/formatter/heading_formatter.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 module Reek
   module Report
-    module HeadingFormatter
+    module Formatter
       #
       # Base class for heading formatters.
       # Is responsible for formatting the heading emitted for each examiner
       #
       # @abstract Override {#show_header?} to implement a heading formatter.
-      class Base
+      class HeadingFormatterBase
         attr_reader :report_formatter
 
         def initialize(report_formatter)
@@ -31,7 +31,7 @@ module Reek
       #
       # Lists out each examiner, even if it has no smell
       #
-      class Verbose < Base
+      class VerboseHeadingFormatter < HeadingFormatterBase
         def show_header?(_examiner)
           true
         end
@@ -40,7 +40,7 @@ module Reek
       #
       # Lists only smelly examiners
       #
-      class Quiet < Base
+      class QuietHeadingFormatter < HeadingFormatterBase
         # :reek:UtilityFunction
         def show_header?(examiner)
           examiner.smelly?
