+++ lib/reek/report/html_report/html_report.html.erb
