+++ spec/reek/report/html_report_spec.rb
@@ -1,6 +1,6 @@
 require_relative '../../spec_helper'
 require_lib 'reek/examiner'
-require_lib 'reek/report/report'
+require_lib 'reek/report/html_report'
 
 RSpec.describe Reek::Report::HTMLReport do
   let(:instance) { described_class.new }
