+++ spec/reek/report/json_report_spec.rb
@@ -1,6 +1,6 @@
 require_relative '../../spec_helper'
 require_lib 'reek/examiner'
-require_lib 'reek/report/report'
+require_lib 'reek/report/json_report'
 require_lib 'reek/report/formatter'
 
 require 'json'
@@ -55,7 +55,7 @@ RSpec.describe Reek::Report::JSONReport do
     end
 
     context 'with link formatter' do
-      let(:options) { { warning_formatter: Reek::Report::WikiLinkWarningFormatter.new } }
+      let(:options) { { warning_formatter: Reek::Report::Formatter::WikiLinkWarningFormatter.new } }
 
       it 'prints documentation links' do
         out = StringIO.new
